RN-342 Refactor post textbox component
